Develop

+++ atlas.sh
@@ -1,8 +1,8 @@
 #!/system/bin/sh
-# version 0.18
+# version 1.0
 
 #Version checks
-Ver55atlas="0.3"
+Ver55atlas="1.0"
 ### add webhook sender?
 
 #Create logfile
@@ -16,15 +16,28 @@ pdconf="/data/data/com.mad.pogodroid/shared_prefs/com.mad.pogodroid_preferences.
 [[ -d /data/data/de.grennith.rgc.remotegpscontroller ]] && ruser=$(ls -la /data/data/de.grennith.rgc.remotegpscontroller/ |head -n2 | tail -n1 | awk '{print $3}')
 rgcconf="/data/data/de.grennith.rgc.remotegpscontroller/shared_prefs/de.grennith.rgc.remotegpscontroller_preferences.xml"
 aconf="/data/local/tmp/atlas_config.json"
-[[ -f /data/local/tmp/aconf_download ]] && aconf_download=$(cat /data/local/tmp/aconf_download | head -n1 )
+aconf_versions="/data/local/aconf_versions"
+[[ -f /data/local/aconf_download ]] && aconf_download=$(grep url /data/local/aconf_download | awk -F "=" '{ print $NF }')
+[[ -f /data/local/aconf_download ]] && aconf_user=$(grep authUser /data/local/aconf_download | awk -F "=" '{ print $NF }')
+[[ -f /data/local/aconf_download ]] && aconf_pass=$(grep authPass /data/local/aconf_download | awk -F "=" '{ print $NF }')
+if [[ -f /data/local/tmp/atlas_config.json ]] ;then
+#  origin=$(grep -w 'deviceName' $aconf | awk -F "\"" '{ print $4 }')
+  origin=$(cat $aconf | tr , '\n' | grep -w 'deviceName' | awk -F "\"" '{ print $4 }')
+else
+  if [[ -f /data/data/de.grennith.rgc.remotegpscontroller/shared_prefs/de.grennith.rgc.remotegpscontroller_preferences.xml ]] ;then
+    origin=$(grep -w 'websocket_origin' $rgcconf | sed -e 's/    <string name="websocket_origin">\(.*\)<\/string>/\1/')
+  else
+    echo "`date +%Y-%m-%d_%T` Cannot find origin, that can't be right" >> $logfile
+  fi
+fi
 
 # stderr to logfile
 exec 2>> $logfile
 
 # add atlas.sh command to log
 echo "" >> $logfile
-echo "`date +%Y-%m-%d_%T` ## Executing $(basename $0) $@" >> $logfile
-echo "`date +%Y-%m-%d_%T` download folder set to $aconf_download" >> $logfile
+echo "`date +%Y-%m-%d_%T` ## Executing $(basename $0) $@" >> $logfile0
+# echo "`date +%Y-%m-%d_%T` download folder set to $aconf_download, user is $aconf_user with pass $aconf_pass" >> $logfile
 
 
 ########## Functions
@@ -42,23 +55,29 @@ esac
 
 install_atlas(){
 
-[[ -f /data/local/tmp/aconf_download ]] && aconf_download=$(cat /data/local/tmp/aconf_download | head -n1 ) && echo "`date +%Y-%m-%d_%T` download folder set to $aconf_download" >> $logfile
-
 # install 55atlas
 mount -o remount,rw /system
-until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/master/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
-  sleep 2
-done
-chmod +x /system/etc/init.d/55atlas
+if [ -f /sdcard/useAconfDevelop ] ;then
+  until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/develop/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
+    sleep 2
+  done
+  chmod +x /system/etc/init.d/55atlas
+  echo "`date +%Y-%m-%d_%T` 55atlas installed, from develop" >> $logfile
+else
+  until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/master/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
+    sleep 2
+  done
+  chmod +x /system/etc/init.d/55atlas
+  echo "`date +%Y-%m-%d_%T` 55atlas installed, from master" >> $logfile
+fi
 mount -o remount,ro /system
-echo "`date +%Y-%m-%d_%T` 55atlas installed" >> $logfile
 
 # get version
-aversion=$(head -2 /data/local/tmp/aconf_versions | grep 'atlas' | awk -F "=" '{ print $NF }')
+aversion=$(grep 'atlas' $aconf_versions | awk -F "=" '{ print $NF }')
 
 # download atlas
 /system/bin/rm -f /sdcard/Download/atlas.apk
-until /system/bin/curl -k -s -L --fail --show-error -o /sdcard/Download/atlas.apk $aconf_download/PokemodAtlas-Public-$aversion.apk || { echo "`date +%Y-%m-%d_%T` Download atlas failed, exit script" >> $logfile ; exit 1; } ;do
+until $download /sdcard/Download/atlas.apk $aconf_download/PokemodAtlas-Public-$aversion.apk || { echo "`date +%Y-%m-%d_%T` Download atlas failed, exit script" >> $logfile ; exit 1; } ;do
   sleep 2
 done
 
@@ -72,6 +91,9 @@ if [ -f "$pdconf" ] ;then
   # disable pd autoupdate
   touch /sdcard/disableautopogodroidupdate
 fi
+
+#disable pogo update by 42mad
+touch /sdcard/disableautopogoupdate
 
 # let us kill pogo as well and clear data
 am force-stop com.nianticlabs.pokemongo
@@ -91,11 +113,10 @@ pm grant com.pokemod.atlas android.permission.WRITE_EXTERNAL_STORAGE
 echo "`date +%Y-%m-%d_%T` atlas granted su and settings set" >> $logfile
 
 # download atlas config file and adjust orgin to rgc setting
-until /system/bin/curl -s -k -L --fail --show-error -o /data/local/tmp/atlas_config.json $aconf_download/atlas_config.json || { echo "`date +%Y-%m-%d_%T` Download atlas config file failed, exit script" >> $logfile ; exit 1; } ;do
+until $download /data/local/tmp/atlas_config.json $aconf_download/atlas_config.json || { echo "`date +%Y-%m-%d_%T` Download atlas config file failed, exit script" >> $logfile ; exit 1; } ;do
   sleep 2
 done
-rgc_origin=$(grep -w 'websocket_origin' $rgcconf | sed -e 's/    <string name="websocket_origin">\(.*\)<\/string>/\1/')
-sed -i 's,dummy,'$rgc_origin',g' /data/local/tmp/atlas_config.json
+sed -i 's,dummy,'$origin',g' $aconf
 
 # check pogo version else remove+install
 downgrade_pogo
@@ -113,14 +134,14 @@ reboot=1
 
 update_all(){
 pinstalled=$(dumpsys package com.nianticlabs.pokemongo | grep versionName | head -n1 | sed 's/ *versionName=//')
-pversions=$(head -2 /data/local/tmp/aconf_versions | grep 'pogo' | awk -F "=" '{ print $NF }')
+pversions=$(grep 'pogo' $aconf_versions | awk -F "=" '{ print $NF }')
 ainstalled=$(dumpsys package com.pokemod.atlas | grep versionName | head -n1 | sed 's/ *versionName=//')
-aversions=$(head -2 /data/local/tmp/aconf_versions | grep 'atlas' | awk -F "=" '{ print $NF }' | awk '{print substr($1,2); }')
+aversions=$(grep 'atlas' $aconf_versions | awk -F "=" '{ print $NF }' | awk '{print substr($1,2); }')
 
 if [ $pinstalled != $pversions ] ;then
   echo "`date +%Y-%m-%d_%T` New pogo version detected, $pinstalled=>$pversions" >> $logfile
   /system/bin/rm -f /sdcard/Download/pogo.apk
-  until /system/bin/curl -s -k -L --fail --show-error -o /sdcard/Download/pogo.apk $aconf_download/pokemongo_$arch\_$pversions.apk || { echo "`date +%Y-%m-%d_%T` Download pogo failed, exit script" >> $logfile ; exit 1; } ;do
+  until $download /sdcard/Download/pogo.apk $aconf_download/pokemongo_$arch\_$pversions.apk || { echo "`date +%Y-%m-%d_%T` Download pogo failed, exit script" >> $logfile ; exit 1; } ;do
     sleep 2
   done
   # set pogo to be installed
@@ -133,7 +154,7 @@ fi
 if [ $ainstalled != $aversions ] ;then
   echo "`date +%Y-%m-%d_%T` New atlas version detected, $ainstalled=>$aversions" >> $logfile
   /system/bin/rm -f /sdcard/Download/atlas.apk
-  until /system/bin/curl -k -s -L --fail --show-error -o /sdcard/Download/atlas.apk $aconf_download/PokemodAtlas-Public-$aversion.apk || { echo "`date +%Y-%m-%d_%T` Download atlas failed, exit script" >> $logfile ; exit 1; } ;do
+  until $download /sdcard/Download/atlas.apk $aconf_download/PokemodAtlas-Public-$aversion.apk || { echo "`date +%Y-%m-%d_%T` Download atlas failed, exit script" >> $logfile ; exit 1; } ;do
     sleep 2
   done
   # set atlas to be installed
@@ -167,7 +188,7 @@ fi
 
 check_rgc(){
 if [ -f "$rgcconf" ] ;then
-  rgccheck=$(head -2 /data/local/tmp/aconf_versions | grep 'rgc' | awk -F "=" '{ print $NF }')
+  rgccheck=$(grep 'rgc' $aconf_versions | awk -F "=" '{ print $NF }')
   rgcstatus=$(grep -w 'boot_startup' $rgcconf | awk -F "\"" '{print tolower($4)}')
   if [[ $rgccheck == "off" ]] && [[ $rgcstatus == "true" ]] ;then
     # disable rgc
@@ -192,15 +213,13 @@ if [ -f "$rgcconf" ] ;then
     echo "`date +%Y-%m-%d_%T` rgc enabled and started" >> $logfile
   fi
 fi
-
-
 }
 
 downgrade_pogo(){
 pinstalled=$(dumpsys package com.nianticlabs.pokemongo | grep versionName | head -n1 | sed 's/ *versionName=//')
-pversions=$(head -2 /data/local/tmp/aconf_versions | grep 'pogo' | awk -F "=" '{ print $NF }')
+pversions=$(grep 'pogo' $aconf_versions | awk -F "=" '{ print $NF }')
 if [ $pinstalled != $pversions ] ;then
-  until /system/bin/curl -s -k -L --fail --show-error -o /sdcard/Download/pogo.apk $aconf_download/pokemongo_$arch\_$pversions.apk || { echo "`date +%Y-%m-%d_%T` Download pogo failed, exit script" >> $logfile ; exit 1; } ;do
+  until $download /sdcard/Download/pogo.apk $aconf_download/pokemongo_$arch\_$pversions.apk || { echo "`date +%Y-%m-%d_%T` Download pogo failed, exit script" >> $logfile ; exit 1; } ;do
     sleep 2
   done
   /system/bin/pm uninstall com.nianticlabs.pokemongo
@@ -220,16 +239,22 @@ until ping -c1 8.8.8.8 >/dev/null 2>/dev/null || ping -c1 1.1.1.1 >/dev/null 2>/
 done
 echo "`date +%Y-%m-%d_%T` Internet connection available" >> $logfile
 
+
 #download latest atlas.sh
 if [[ $(basename $0) != "atlas_new.sh" ]] ;then
   mount -o remount,rw /system
   oldsh=$(head -2 /system/bin/atlas.sh | grep '# version' | awk '{ print $NF }')
-
-  until /system/bin/curl -s -k -L --fail --show-error -o /system/bin/atlas_new.sh https://raw.githubusercontent.com/dkmur/aconf/master/atlas.sh || { echo "`date +%Y-%m-%d_%T` Download atlas.sh failed, exit script" >> $logfile ; exit 1; } ;do
-    sleep 2
-  done
-  chmod +x /system/bin/atlas_new.sh
-
+  if [ -f /sdcard/useAconfDevelop ] ;then
+    until /system/bin/curl -s -k -L --fail --show-error -o /system/bin/atlas_new.sh https://raw.githubusercontent.com/dkmur/aconf/develop/atlas.sh || { echo "`date +%Y-%m-%d_%T` Download atlas.sh failed, exit script" >> $logfile ; exit 1; } ;do
+      sleep 2
+    done
+    chmod +x /system/bin/atlas_new.sh
+  else
+    until /system/bin/curl -s -k -L --fail --show-error -o /system/bin/atlas_new.sh https://raw.githubusercontent.com/dkmur/aconf/master/atlas.sh || { echo "`date +%Y-%m-%d_%T` Download atlas.sh failed, exit script" >> $logfile ; exit 1; } ;do
+      sleep 2
+    done
+    chmod +x /system/bin/atlas_new.sh
+  fi
   newsh=$(head -2 /system/bin/atlas_new.sh | grep '# version' | awk '{ print $NF }')
   if [[ $oldsh != $newsh ]] ;then
     echo "`date +%Y-%m-%d_%T` atlas.sh $oldsh=>$newsh, restarting script" >> $logfile
@@ -241,22 +266,39 @@ if [[ $(basename $0) != "atlas_new.sh" ]] ;then
   fi
 fi
 
+
 #update 55atlas if needed
 if [[ $(basename $0) = "atlas_new.sh" ]] ;then
   old55=$(head -2 /system/etc/init.d/55atlas | grep '# version' | awk '{ print $NF }')
   if [ $Ver55atlas != $old55 ] ;then
     mount -o remount,rw /system
-    until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/master/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
-      sleep 2
-    done
-    chmod +x /system/etc/init.d/55atlas
+    if [ -f /sdcard/useAconfDevelop ] ;then
+      until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/develop/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
+        sleep 2
+      done
+      chmod +x /system/etc/init.d/55atlas
+    else
+      until /system/bin/curl -s -k -L --fail --show-error -o /system/etc/init.d/55atlas https://raw.githubusercontent.com/dkmur/aconf/master/55atlas || { echo "`date +%Y-%m-%d_%T` Download 55atlas failed, exit script" >> $logfile ; exit 1; } ;do
+        sleep 2
+      done
+      chmod +x /system/etc/init.d/55atlas
+    fi
     mount -o remount,ro /system
     new55=$(head -2 /system/etc/init.d/55atlas | grep '# version' | awk '{ print $NF }')
     echo "`date +%Y-%m-%d_%T` 55atlas $old55=>$new55" >> $logfile
   fi
 fi
 
-### verify download folder ??
+# verify download credential file and set download
+if [[ ! -f /data/local/aconf_download ]] ;then
+  echo "`date +%Y-%m-%d_%T` File /data/local/aconf_download not found, exit script" >> $logfile && exit 1
+else
+  if [[ $aconf_user == "" ]] ;then
+    download="/system/bin/curl -s -k -L --fail --show-error -o"
+  else
+    download="/system/bin/curl -s -k -L --fail --show-error --user $aconf_user:$aconf_pass -o"
+  fi
+fi
 
 # prevent amconf causing reboot loop. Add bypass ??
 if [ $(cat /sdcard/aconf.log | grep `date +%Y-%m-%d` | grep rebooted | wc -l) -gt 20 ] ;then
@@ -264,10 +306,29 @@ if [ $(cat /sdcard/aconf.log | grep `date +%Y-%m-%d` | grep rebooted | wc -l) -g
   exit 1
 fi
 
+# set hostname = origin, wait till next reboot for it to take effect
+if [[ $origin != "" ]] ;then
+  if [ $(cat /system/build.prop | grep net.hostname | wc -l) = 0 ]; then
+    mount -o remount,rw /system
+    echo "`date +%Y-%m-%d_%T` No hostname set, setting it to $origin" >> $logfile
+    echo "net.hostname=$origin" >> /system/build.prop
+    mount -o remount,ro /system
+  else
+    hostname=$(grep net.hostname /system/build.prop | awk 'BEGIN { FS = "=" } ; { print $2 }')
+    if [[ $hostname != $origin ]] ;then
+      mount -o remount,rw /system
+      echo "`date +%Y-%m-%d_%T` Changing hostname, from $hostname to $origin" >> $logfile
+      sed -i -e "s/^net.hostname=.*/net.hostname=$origin/g" /system/build.prop
+      mount -o remount,ro /system
+    fi
+  fi
+fi
+
 # download latest version file
-until /system/bin/curl -s -k -L --fail --show-error -o /data/local/tmp/aconf_versions $aconf_download/versions || { echo "`date +%Y-%m-%d_%T` Download atlas version file failed, exit script" >> $logfile ; exit 1; } ;do
+until $download $aconf_versions $aconf_download/versions || { echo "`date +%Y-%m-%d_%T` Download atlas versions file failed, exit script" >> $logfile ; exit 1; } ;do
   sleep 2
 done
+echo "`date +%Y-%m-%d_%T` Downloaded latest versions file"  >> $logfile
 
 # check rgc enable/disable
 check_rgc
@@ -277,7 +338,8 @@ for i in "$@" ;do
  -ia) install_atlas ;;
  -ua) update_all ;;
  -dp) downgrade_pogo;;
-# consider adding: downgrade pogo, downgrade atlas, update atlas config file, update donwload link
+ -cr) check_rgc;;
+# consider adding: downgrade atlas, update atlas config file, update donwload link
  esac
 done
